+++ src/grammar/ast/action/mod.rs
@@ -1,6 +1,7 @@
 pub mod discard;
 pub mod fileinto;
 pub mod keep;
+pub mod notify;
 pub mod redirect;
 pub mod reject;
 pub mod vacation;
@@ -8,6 +9,7 @@ pub mod vacation;
 pub use discard::ActionDiscard;
 pub use fileinto::ActionFileinto;
 pub use keep::ActionKeep;
+pub use notify::ActionNotify;
 pub use redirect::ActionRedirect;
 pub use reject::ActionReject;
 pub use vacation::ActionVacation;