+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.1.3] - 2023-10-24
+
+### Fixed
+
+- Fix comment position in `if` statement
+- Fix comment in the same line as `if` statement
+
+### Added
+
+- Implement `notify` action (RFC 5429)
+
 ## [0.1.2] - 2023-10-20
 
 - Second release